+++ configuration/data/config/init-config.sh
@@ -4,16 +4,19 @@
 ##   - query fields (including boost value)
 ##   - response fields
 ##   - highlight fields
-##   - facet fields
+##   - aggregation fields
 ##   - filter fields
 ##
-## You can use first commandline parameter to change base URL of system API call (/v1/rest/... is appended automatically to this base URL)
+## You can use first commandline parameter to change base URL of system API call (/v2/rest/... is appended automatically to this base URL)
 ## You can use second commandline parameter to change system username
 ## You can use third commandline parameter to change system password
 
 clear
 
 sysurl=https://dcp-jbossorgdev.rhcloud.com
+if [ -n "${OPENSHIFT_JBOSSEAP_IP}" ]; then
+  sysurl=http://${OPENSHIFT_JBOSSEAP_IP}:8080
+fi
 if [ -n "$1" ]; then
   sysurl=$1
 fi
@@ -29,7 +32,7 @@ if [ -n "$3" ]; then
 fi
 
 outputfile=output.txt
-sysapi=${sysurl}/v1/rest/config/
+sysapi=${sysurl}/v2/rest/config/
 
 echo "Pushing configuration documents to system API via ${sysapi}"
 echo -n "" > $outputfile